backend/requirements.txt
# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# FastAPI Framework (New - for abandoned vehicle detection)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Computer Vision & AI
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

# Image Processing
Pillow==10.1.0
numpy==1.24.4
scikit-image==0.21.0

# Machine Learning & Analysis
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Geospatial
geopandas==0.14.1
shapely==2.0.2
folium==0.15.0
pyproj==3.6.1
fiona==1.9.5

# HTTP & API
requests==2.31.0
urllib3==2.1.0

# Configuration & Environment
python-dotenv==1.0.0

# Date/Time Processing
python-dateutil==2.8.2

# JSON Processing
jsonschema==4.20.0

# Satellite Data Processing
rasterio==1.3.9
gdal==3.8.0

# Parallel Processing
joblib==1.3.2

# Memory Optimization
psutil==5.9.6

# PDF Processing
pdf2image==1.16.3
PyMuPDF==1.23.8

# Development & Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

